file util and audio entry
update file handling a little bit when creating temp file names. it will
now be faster hopefully and will do a better check to see if the file
exists in the directory given.

small changes to how files are handled for audio entries. the error
handling was updated and should now be a bit easier to follow.