entries update and response schema's(?)
updated entries table to include created, updated, and deleted timestamps.
updated the response jsons to use a modified version that makes some more sense. it will not include entry ids for nested data and renamed some of the keys to be more concise.
for the list request it will no longer return the all data for each entry and just include some information. text and audio will now just be counts of how many values were found. also added in counts for future updates of files and video.